% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_lists_to_df.R
\name{flatten_lists_to_df}
\alias{flatten_lists_to_df}
\title{A wrapper to flatten a list of unpredictably nested lists into a single
long dataframe.}
\usage{
flatten_lists_to_df(list)
}
\arguments{
\item{list}{A list of nested lists.}
}
\value{
Returns a long-format dataframe with each row being a
  source-path-order-type-value observation derived from the list of nested
  lists.
}
\description{
\code{flatten_lists_to_df} uses \code{flatten_list_to_long_df} to flatten
each unpredictably nested sublist into a long dataframe associated with
the sublist name (or order, if names are not provided).
}
\details{
It then combines these lists into a single long list.
}
\examples{
#' # Load the nested Steam API data.
data(steam_app_list)

# The Steam API data is a list of nested lists, with each nested list
# representing a Steam app (e.g., PC game).
df_demo <- flatten_lists_to_df(steam_app_list[1:5])

}

