% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_steam_app.R
\name{get_steam_app}
\alias{get_steam_app}
\title{Acquire the raw data for a single Steam app.}
\usage{
get_steam_app(appid)
}
\arguments{
\item{appid}{A steam app ID. If integer or numeric, it will be converted
to character.}
}
\value{
The function will return a nested list of app features, converted
 from the JSON returned by the API.
}
\description{
This function interacts with the Steam API to acquire the list of features
available for the app.
}
\details{
\code{get_app} can be used on its own but was primarily designed as a helper
function for \code{get_steam_apps}. That function iterates over a list of
Steam app IDs and uses \code{get_app} to get the features for each app.
}
\examples{
\dontrun{
# grab the IDs and names for all apps currently listed on the Steam API
current_apps <- get_steam_app_ids()

# grab the data for one of the apps
single_app <- get_steam_app(current_apps[500])
}

}

