% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_steam_app_ids.R
\name{get_steam_app_ids}
\alias{get_steam_app_ids}
\title{Get a dataframe of current Steam app ids and names.}
\usage{
get_steam_app_ids()
}
\value{
The function returns a dataframe of the appids and names for all
 current Steam apps (as reported to the Steam API).
}
\description{
The Steam API restricts how much detailed app data you can grab at once and
in a given window of time. However, you are allowed to grab the index of app
ids and names in a single query.
}
\details{
This is a convenience function to handle the index query and to process the
results into a dataframe for easy manipulation in R.

The dataframe is useful for constructing the richer dataset of Steam app
data and assessing for differences between a working dataset and the
data now available on the API.
}
\examples{
\dontrun{
# grab the IDs and names for all apps currently listed on the Steam API
current_apps <- get_steam_app_ids()
}

}

