% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_steam_apps.R
\name{get_steam_apps}
\alias{get_steam_apps}
\title{Acquire raw Steam app list and features.}
\usage{
get_steam_apps(appids, start_delay = 240)
}
\arguments{
\item{appids}{A character vector of Steam app ids. This is most easily
obtained from the dataframe produced by \code{get_steam_app_ids}.}

\item{start_delay}{The function will attempt to measure how long it needs
to wait between calls to the Steam API to avoid timeout requests. This
tells the function what value to start with (in seconds). Too small and
early tests will have a number of denied connections. Too large and the
function will wait longer than it has to between requests.}
}
\value{
The function will return a nested list, converted from the JSON
 the Steam API initially provides.
}
\description{
This function interacts with the Steam API to acquire a list of Steam apps
and the features the API naturally associates with each app.
}
\details{
When cleaned, this list will provide the sample of video game names we will
attach protagonist variables to. The list will also provide several other
game-specific variables of interest (e.g., release date, developer).

In its raw format, the list has a number of non-game apps (e.g., Steam
server apps, collections of games, game videos). It is also based on a
JSON query, and so has nesting that is problematic for clean conversion
to a dataframe. In other words, it will require cleaning (handled by
other functions) to make it a useful dataframe to work with and build on.
}
\section{Run with caution}{

This function is included in large part for documentation purposes - so that
it is clear how our data were sampled and processed.

You are advised \strong{not to run} the function unless you are certain
you want to build the most current version of the Steam API data. Remember
that the vgpd package comes with dated versions of the raw and processed
Steam data for your use.

Additional warnings:
\itemize{
 \item This function requires an active and stable internet connection.
 \item Due to the nature of the Steam API, the function deals with thresholds
     on queries-per-five-minutes (200) and how many games can be grabbed
     per query (1). As a result, the function takes a very long time to
     complete (e.g., the initial run took 8 hours).
}
}
\examples{
\dontrun{
# extract the app IDs from the id-name dataframe produced here
current_apps <- get_steam_app_ids()

# get app data for all the app IDs using the default delay as a good
# starting point
raw_app_data <- get_steam_apps(current_apps$appid)
}

}

