% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgpd.R
\docType{package}
\name{vgpd}
\alias{vgpd}
\alias{vgpd-package}
\title{vgpd: Data and functions for exploring video game protagonist demographics.}
\description{
The vgpd package contains a sample dataset of video games and their
protagonists, functions for reproducing and interacting with the sample
dataset, and documentation discussing the work done by the package authors
to produce and use the data.
}
\section{Warning}{

Please see the package README and data-focused vignettes for details on
how/why the data were sampled and suggested limitations on data use
and interpretation.
}

